Separate opcodes into modules
